test: 사업자등록증 상태 조회 테스트 로직 통합
